+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 2.2.0
 
 - Fixed style target string field overlapping delete button
+- Moved title text field to advanced options